Move <script>, Remove TogetherJS
(index.html)
Remove everything related to Together.js because it's causing problems.
Why not get rid of Flattr while we're at it, this page is going to be
gutted anyway.
As per the default Content Security Policy, inline JavaScript will not
be executed. Therefore, the script in this document (index.html) was
extracted to an internal.js